requirements.txt
gradio>=4.43.0
huggingface-hub>=0.27.0
python-dotenv>=1.0.0
